# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "API documentation for my Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/company-news": {
            "get": {
                "tags": [
                    "Company News"
                ],
                "summary": "Fetch all news",
                "operationId": "dcb016f3bbe182b6e88f93a20bccebc2",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Company News"
                ],
                "summary": "Store new news",
                "operationId": "b44d37c522d13809c76f486dbd453c65",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "companyNewTitle",
                                    "companyNewUrl",
                                    "companyNewImageUrl"
                                ],
                                "properties": {
                                    "companyNewTitle": {
                                        "description": "Title of the news",
                                        "type": "string"
                                    },
                                    "companyNewUrl": {
                                        "description": "URL of the news",
                                        "type": "string"
                                    },
                                    "companyNewImageUrl": {
                                        "description": "Image URL of the news",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "News created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-news/{id}": {
            "get": {
                "tags": [
                    "Company News"
                ],
                "summary": "Show a single news item",
                "operationId": "459cd2521987eab6cb9fd21bba990257",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "News not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Company News"
                ],
                "summary": "Update a news item",
                "operationId": "17dd9972d41b52e75728a8222cf3f902",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "companyNewTitle": {
                                        "description": "Title of the news",
                                        "type": "string"
                                    },
                                    "companyNewUrl": {
                                        "description": "URL of the news",
                                        "type": "string"
                                    },
                                    "companyNewImageUrl": {
                                        "description": "Image URL of the news",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "News updated successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "404": {
                        "description": "News not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Company News"
                ],
                "summary": "Delete a news item",
                "operationId": "062a9990b832b3c0c485b49f703ac92c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News deleted successfully"
                    },
                    "404": {
                        "description": "News not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}